refactor: update import paths and remove unnecessary code
