+++ app/api/users/[userId]/route.ts
@@ -1,6 +1,7 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { z } from 'zod';
-import { HttpCodes, messages } from '@/constants/api';
+import { HttpCodes } from '@/constants/api';
+import { messages } from '@/constants/messages';
 import { checkAdminAccess, createResponseWithError, getUserAndValidate, responseInit } from '@/lib/api-utils';
 import { ApiRequestError, ApiRequestZodError } from '@/lib/error';
 import { prismaClient } from '@/lib/prisma';
