+++ app/not-found.tsx
@@ -1,7 +1,8 @@
 'use client';
 
 import ErrorPage from 'next/error';
-import { HttpCodes, messages } from 'constants/api';
+import { HttpCodes } from 'constants/api';
+import { messages } from '@/constants/messages';
 
 export default function NotFound() {
   return <ErrorPage statusCode={Number(HttpCodes.NotFound)} title={messages.pageNotFound} />;
