+++ constants/messages.ts
@@ -1,0 +1,15 @@
+export const messages = {
+  invalidAccess: '不正なアクセスです',
+  invalidToken: 'トークンの検証に失敗しました',
+  userNotFound: 'ユーザが見つかりません',
+  sessionNotFound: 'セッションが見つかりません',
+  failedSession: 'セッションが不正です',
+  needLogin: 'ログインしてください',
+  adminOnly: '管理者権限が必要です',
+  forbidden: 'アクセス権がありません',
+  notFound: '見つかりません',
+  unknownError: '不具合が発生しました',
+  pageNotFound: 'ページが見つかりません',
+  internalServerError: 'Internal Server Error.',
+  cantModifyMyself: '自分の権限は修正できません',
+};
